change the type of time in sleep() into unsigned long #18

+++ src/Zackernel.h
@@ -39,7 +39,7 @@ void printQueue();
 
 void dispatch();
 
-void sleep(int time, vl::Func<void(void)> func);
+void sleep(unsigned long time, vl::Func<void(void)> func);
 
 void fork(vl::Func<void(void)> func1, vl::Func<void(void)> func2);
